+++ extension/src/ALObject/Maps.ts
@@ -45,6 +45,7 @@ export const multiLanguageTypeMap = new Map<string, MultiLanguageType>([
 
 export const alPropertyTypeMap = new Map<string, ALPropertyType>([
   ["sourcetable", ALPropertyType.sourceTable],
+  ["tableno", ALPropertyType.tableNo],
   ["tabletype", ALPropertyType.tableType],
   ["pagetype", ALPropertyType.pageType],
   ["querytype", ALPropertyType.queryType],